streamlit/requirements.txt
streamlit
psycopg2-binary
requests
pandas
google-generativeai
Pillow